Updated app bar and more info

+++ AniGuide/app/src/main/java/com/example/aniguide/api/SearchMeta.kt
@@ -22,7 +22,7 @@ data class SearchResults (
     @SerializedName("backdrop_path") val backdrop_path : String,
     @SerializedName("original_language") val original_language : String,
     @SerializedName("id") val id : Int,
-    @SerializedName("vote_average") val vote_average : Int,
+    @SerializedName("vote_average") val vote_average : Float,
     @SerializedName("overview") val overview : String,
     @SerializedName("poster_path") val poster_path : String
 )
